requirements.txt
ag-grid-django==0.0.7.8
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.9.1
attrs==25.3.0
bcrypt==4.3.0
billiard==4.2.1
black==25.1.0
boto3==1.40.1
botocore==1.40.1
build==1.3.0
celery==5.5.3
certifi==2025.7.14
cffi==1.17.1
cfgv==3.4.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.2
click==8.2.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorclass==2.2.2
cron-descriptor==1.4.5
cryptography==45.0.5
diff-match-patch==20241021
distlib==0.4.0
Django==5.2.4
django-browser-reload==1.18.0
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-debug-toolbar==6.0.0
django-environ==0.12.0
django-extensions==4.1
django-import-export==4.3.9
django-ninja==1.4.3
django-storages==1.14.6
django-timezone-field==7.1
django-unfold-patrick==0.66.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
docopt==0.6.2
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.9.1
et_xmlfile==2.0.0
Faker==37.5.3
filelock==3.18.0
frozenlist==1.7.0
http_ece==1.2.1
identify==2.6.12
idna==3.10
inflection==0.5.1
isort==6.0.1
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
latest==0.6.0
mccabe==0.7.0
msgpack==1.1.1
multidict==6.6.4
mypy_extensions==1.1.0
nodeenv==1.9.1
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
paramiko==3.5.1
pathspec==0.12.1
pillow==11.3.0
pip-review==1.3.0
pip-upgrader==1.4.15
platformdirs==4.3.8
pre_commit==4.2.0
prompt_toolkit==3.0.51
propcache==0.3.2
psycopg2==2.9.10
py-vapid==1.9.2
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydotplus==2.0.2
pyflakes==3.4.0
PyJWT==2.10.1
PyNaCl==1.5.0
pyparsing==3.2.3
pyproject_hooks==1.2.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.27.1
s3transfer==0.13.1
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
terminaltables==3.1.10
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
virtualenv==20.32.0
wcwidth==0.2.13
yarl==1.20.1
